# api/docs/openapi.yaml
openapi: 3.0.0
info:
    title: 'Zyvo API'
    version: 1.0.0
    description: "\n        A secure Node.js API server built with Express.js that implements \n        security best practices including CORS, CSRF protection, rate limiting, \n        and comprehensive error handling.\n        \n        ## Features\n        - JWT Authentication\n        - Role-based Access Control\n        - Rate Limiting\n        - CSRF Protection\n        - Input Validation\n        - Comprehensive Error Handling\n        \n        ## Security\n        This API implements multiple layers of security:\n        - CORS with configurable origins\n        - CSRF token validation\n        - Rate limiting per IP\n        - Secure HTTP headers via Helmet.js\n        - Input sanitization and validation\n      "
    contact:
        name: 'Zyvo API Support'
        email: support@zyvo.com
        url: 'https://zyvo.com'
    license:
        name: ISC
        url: 'https://opensource.org/licenses/ISC'
servers:
    -
        url: 'http://localhost:3000'
        description: 'Development server'
    -
        url: 'https://api.zyvo.com'
        description: 'Production server'
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: 'JWT Bearer token authentication'
        csrfToken:
            type: apiKey
            in: header
            name: X-CSRF-Token
            description: 'CSRF token for protection against CSRF attacks'
        BearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
...
paths:
...
    /api/v1/inventory:
        post:
            summary: 'Create a new inventory record'
            tags: [Inventory]
            security: [{adminAuth: []}]
            requestBody: {required: true, content: {application/json: {schema: {type: object, required: [product_variant_id], properties: {product_variant_id: {type: string, description: 'ID of the base unit product variant'}, stock_quantity: {type: number, minimum: 0, default: 0, description: 'Initial stock quantity'}, min_stock_level: {type: number, minimum: 0, default: 0, description: 'Minimum stock threshold'}, location: {type: string, maxLength: 200, description: 'Physical storage location'}, notes: {type: string, maxLength: 1000, description: 'Internal notes'}}}}}}
            responses: {'201': {description: 'Inventory record created successfully'}, '400': {description: 'Validation error or pack variant provided'}, '401': {description: 'Unauthorized - admin access required'}, '500': {description: 'Internal server error'}}
        get:
            summary: 'Get all inventory records with pagination and filtering'
            tags: [Inventory]
            security: [{adminAuth: []}]
            parameters: [{in: query, name: page, schema: {type: integer, minimum: 1, default: 1}, description: 'Page number for pagination'}, {in: query, name: limit, schema: {type: integer, minimum: 1, maximum: 100, default: 10}, description: 'Number of items per page'}, {in: query, name: is_active, schema: {type: boolean}, description: 'Filter by active status'}, {in: query, name: stock_status, schema: {type: string, enum: [out_of_stock, low_stock, in_stock]}, description: 'Filter by stock status'}, {in: query, name: location, schema: {type: string}, description: 'Filter by location (partial match)'}, {in: query, name: product_id, schema: {type: string}, description: 'Filter by product ID'}, {in: query, name: search, schema: {type: string}, description: 'Search across SKU, location, and notes'}, {in: query, name: include_computed_packs, schema: {type: boolean, default: false}, description: 'Include computed stock for pack variants'}, {in: query, name: sort_by, schema: {type: string, enum: [createdAt, updatedAt, stock_quantity, min_stock_level, last_restock_date], default: createdAt}, description: 'Field to sort by'}, {in: query, name: sort_order, schema: {type: string, enum: [asc, desc], default: desc}, description: 'Sort order'}]
            responses: {'200': {description: 'Inventory records retrieved successfully'}, '401': {description: 'Unauthorized - admin access required'}, '500': {description: 'Internal server error'}}
    '/api/v1/inventory/variant/{productVariantId}':
        get:
            summary: 'Get inventory by product variant ID (computed stock)'
            tags: [Inventory]
            security: [{adminAuth: []}]
            parameters: [{in: path, name: productVariantId, required: true, schema: {type: string}, description: 'Product variant ID (base unit or pack)'}]
            responses: {'200': {description: 'Inventory record retrieved successfully with computed stock'}, '400': {description: 'Invalid product variant ID'}, '401': {description: 'Unauthorized - admin access required'}, '404': {description: 'Inventory record not found'}, '500': {description: 'Internal server error'}}
    '/api/v1/inventory/{id}':
        patch:
            summary: 'Update inventory record'
            tags: [Inventory]
            security: [{adminAuth: []}]
            parameters: [{in: path, name: id, required: true, schema: {type: string}, description: 'Inventory record ID'}]
            requestBody: {required: true, content: {application/json: {schema: {type: object, properties: {stock_quantity: {type: number, minimum: 0, description: 'Updated stock quantity'}, min_stock_level: {type: number, minimum: 0, description: 'Updated minimum stock level'}, location: {type: string, maxLength: 200, description: 'Updated storage location'}, notes: {type: string, maxLength: 1000, description: 'Updated notes'}, is_active: {type: boolean, description: 'Updated active status'}}}}}}
            responses: {'200': {description: 'Inventory record updated successfully'}, '400': {description: 'Validation error or invalid inventory ID'}, '401': {description: 'Unauthorized - admin access required'}, '404': {description: 'Inventory record not found'}, '500': {description: 'Internal server error'}}
        delete:
            summary: 'Delete inventory record (soft delete)'
            tags: [Inventory]
            security: [{adminAuth: []}]
            parameters: [{in: path, name: id, required: true, schema: {type: string}, description: 'Inventory record ID'}]
            responses: {'204': {description: 'Inventory record deleted successfully'}, '400': {description: 'Invalid inventory ID'}, '401': {description: 'Unauthorized - admin access required'}, '404': {description: 'Inventory record not found'}, '500': {description: 'Internal server error'}}
...